fix(probes): update liveness and readiness probes to use TCP socket checks
This commit modifies the liveness and readiness probes in the Helm chart to utilize TCP socket checks instead of HTTP GET requests. This change enhances the reliability of health checks for the external-node service, ensuring accurate detection of service availability. The initial delay settings for both probes have also been specified in the values.yaml file.

Signed-off-by: Roderik van der Veer <roderik@settlemint.com>